Reagan's modification with usecase classes and controller!
